Change r1 in four_arg_syscall to a u32.
I mistakenly kept it as a `usize` in my previous commit because memop passes pointers using r1, but four_arg_syscall is not used to call memop. Therefore the second argument can always be a u32.

+++ core/platform/src/raw_syscalls.rs
@@ -116,7 +116,7 @@ pub trait RawSyscalls {
     /// depending on the system call class.
     unsafe fn four_arg_syscall(
         r0: u32,
-        r1: usize,
+        r1: u32,
         r2: usize,
         r3: usize,
         class: u8,
